requirements.txt
astroid==2.12.13
autopep8==2.0.0
bleach==6.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
dill==0.3.6
docutils==0.19
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
isort==5.10.1
jaraco.classes==3.2.3
keyring==23.13.1
lazy-object-proxy==1.8.0
markdown-it-py==2.2.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.0.0
pkginfo==1.9.6
platformdirs==2.5.4
pycodestyle==2.9.1
Pygments==2.14.0
pylint==2.15.6
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.3.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
twine==4.0.2
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.12
webencodings==0.5.1
wrapt==1.14.1
zipp==3.13.0

coinbaseadvanced~=0.1.1
setuptools~=57.4.0
python-dotenv~=1.0.0